Get Java project to compile and run tests
