+++ java/videostore/src/main/java/com/example/videostore/Rental.java
@@ -1,0 +1,20 @@
+package com.example.videostore;
+
+public class Rental
+{
+	public Rental (Movie movie, int daysRented) {
+		this.movie 		= movie;
+		this.daysRented = daysRented;
+	}
+
+	public int getDaysRented () {
+		return daysRented;
+	}
+
+	public Movie getMovie () {
+		return movie;
+	}
+
+	private Movie movie;
+	private int daysRented;
+}
